+++ core/run/test/Jobs.cpp
@@ -149,7 +149,7 @@ public:
 
 static void registerJobs() {
     static SphJob sSph("");
-    static CollisionGeometrySetup sSetup("");
+    static CollisionGeometrySetupJob sSetup("");
     static MonolithicBodyIc sIc("");
     static SaveFileJob sIo("");
     static BlockJob sBlock("");